refactor: refactor files on aside menu and add custon scrollbar
