+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "workbench.colorTheme": "One Dark Pro Darker"
+  "workbench.colorTheme": "Bio Dark"
 }
