+++ components/AsideMenu/styles.module.css
@@ -1,8 +1,3 @@
-.anchor,
-.menuItemLinkInactive {
-  color: var(--white);
-}
-
 .asideMenu {
   height: 100vh;
   padding-top: 65px;
@@ -16,87 +11,6 @@
   width: 220px;
 }
 
-.ulActive {
-  margin: 0;
-  padding: 0;
-  padding-top: 0.4rem;
-  display: flex;
-  flex-direction: column;
-}
-
-.menuItemActive {
-  list-style: none;
-  display: flex;
-  justify-content: center;
-}
-
-.menuItemLinkActive {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  font-size: 0.87rem;
-  line-height: 20px;
-  font-family: 'Roboto Medium';
-  gap: 1.4rem;
-  padding-left: 1.6rem;
-  color: var(--white);
-}
-
-.menuItemLinkActive:hover {
-  background-color: var(--state-bg-color-dark);
-}
-
-.buttonActive,
-.buttonInactive {
-  background-color: var(--state-bg-color-dark);
-}
-
-/* INACTIVE MENU: */
 .inactiveMenu {
   width: 73px;
-}
-
-.ul {
-  margin: 0;
-  padding: 0;
-  padding-top: 1rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.showScrollUl {
-  overflow-y: scroll;
-}
-
-.menuItem {
-  list-style: none;
-  display: flex;
-  gap: 1rem;
-}
-
-.menuItemLinkInactive {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  font-size: 0.65rem;
-  width: 100%;
-  padding: 14px 0;
-}
-
-.menuItemLinkInactive:hover {
-  background-color: var(--state-bg-color-dark);
-}
-
-.labelInactive {
-  margin: 0;
-  margin-top: 10px;
-}
-
-/* Línea separadora */
-.separator {
-  width: 95%;
-  border-bottom: 1px solid var(--separator);
-  margin: 8px 0;
 }
