+++ styles/globals.css
@@ -37,3 +37,20 @@ a {
   font-family: 'Roboto';
   src: url('/fonts/Roboto.woff2') format('woff2');
 }
+
+body::-webkit-scrollbar {
+  width: 0.5vw;
+}
+
+body::-webkit-scrollbar-button {
+  display: none;
+}
+
+body::-webkit-scrollbar-track {
+  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
+}
+
+body::-webkit-scrollbar-thumb {
+  background-color: darkgrey;
+  border-radius: 10px;
+}
